requirements.txt


streamlit>=1.25.0
pillow>=10.0.0
easyocr>=1.6.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
faster-whisper>=0.6.0
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=1.0.0
google-generativeai>=0.4.1

